# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEORx API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List audit jobs
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of audit jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Create a new audit job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuditBody"
      responses:
        "201":
          description: Created audit job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audit"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /audits/{id}:
    get:
      operationId: getAudit
      tags: [audits]
      summary: Get audit job by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Audit job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAudit
      tags: [audits]
      summary: Delete audit job
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /audits/{id}/issues:
    get:
      operationId: listAuditIssues
      tags: [issues]
      summary: List issues for an audit
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, approved, dismissed, fixed]
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditIssue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ai-providers:
    get:
      operationId: listAiProviders
      tags: [ai-providers]
      summary: List AI provider configurations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of AI providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAiProvider
      tags: [ai-providers]
      summary: Add an AI provider configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiProviderBody"
      responses:
        "201":
          description: Created AI provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /ai-providers/{id}:
    put:
      operationId: updateAiProvider
      tags: [ai-providers]
      summary: Update AI provider configuration
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiProviderBody"
      responses:
        "200":
          description: Updated AI provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAiProvider
      tags: [ai-providers]
      summary: Delete AI provider configuration
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [superadmin, admin, agency, client, viewer]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - role
        - createdAt
        - updatedAt
    UpdateUserProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
...
    Audit:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        seoScore:
          type: number
          nullable: true
        issueCount:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - url
        - status
        - issueCount
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - createdAt
        - updatedAt
    AuditDetail:
      allOf:
        - $ref: "#/components/schemas/Audit"
        - type: object
          properties:
            crawledPages:
              type: integer
            scanDurationMs:
              type: integer
              nullable: true
            aiProviderUsed:
              type: string
              nullable: true
            pageSpeedScore:
              type: number
              nullable: true
            issues:
              type: array
              items:
                $ref: "#/components/schemas/AuditIssue"
    AuditList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Audit"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    CreateAuditBody:
      type: object
      properties:
        clientId:
          type: string
        url:
          type: string
          format: uri
        maxPages:
          type: integer
          default: 100
        includePageSpeed:
          type: boolean
          default: false
        aiProviderId:
          type: string
          nullable: true
      required:
        - clientId
        - url
    AuditIssue:
      type: object
      properties:
        id:
          type: string
        auditId:
          type: string
        url:
          type: string
        category:
          type: string
          enum: [meta, content, performance, links, structured_data, mobile, security, crawlability]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: string
        recommendation:
          type: string
          nullable: true
        aiRecommendation:
          type: string
          nullable: true
        priorityScore:
          type: number
        status:
          type: string
          enum: [open, approved, dismissed, fixed]
        approvedBy:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        affectedElement:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auditId
        - url
        - category
        - severity
        - title
        - description
        - priorityScore
        - status
        - createdAt
...
    AiProvider:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, gemini, ollama, custom]
        model:
          type: string
        baseUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - name
        - provider
        - model
        - isActive
        - isDefault
        - createdAt
        - updatedAt
    CreateAiProviderBody:
      type: object
      properties:
        orgId:
          type: string
        name:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, gemini, ollama, custom]
        model:
          type: string
        apiKey:
          type: string
          nullable: true
        baseUrl:
          type: string
          nullable: true
        isDefault:
          type: boolean
          default: false
      required:
        - orgId
        - name
        - provider
        - model
    UpdateAiProviderBody:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        apiKey:
          type: string
          nullable: true
        baseUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isDefault:
          type: boolean
...